requirements.txt
neo4j
openai
psycopg2-binary
fastmcp